UBER-148: Move ModeSelector and labels from tracker to use in other plugins
