+++ plugins/tracker-resources/src/components/issues/IssuesHeader.svelte
@@ -1,12 +1,11 @@
 <script lang="ts">
   import { Ref, Space } from '@hcengineering/core'
-  import { TabList, SearchEdit } from '@hcengineering/ui'
+  import { TabList, SearchEdit, IModeSelector } from '@hcengineering/ui'
   import { Viewlet } from '@hcengineering/view'
   import { FilterButton, setActiveViewletId } from '@hcengineering/view-resources'
   import tracker from '../../plugin'
   import { WithLookup } from '@hcengineering/core'
-  import ModeSelector from '../ModeSelector.svelte'
-  import { IModeSelector } from '../../utils'
+  import ModeSelector from '@hcengineering/ui/src/components/ModeSelector.svelte'
 
   export let space: Ref<Space> | undefined = undefined
   export let viewlet: WithLookup<Viewlet> | undefined
