+++ plugins/view-assets/lang/ru.json
@@ -31,6 +31,9 @@
     "SelectItem": "Выбрать",
     "SelectItemAll": "Выбрать все",
     "SelectItemNone": "Снять все выделения",
+    "Assigned": "Назначенные",
+    "Created": "Созданные",
+    "Subscribed": "Отслеживаемые",
     "ShowPreview": "Предпросмотр документа",
     "ShowActions": "Показать действия",
     "RestoreDefaults": "По умолчанию",
